pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitast"
version = "0.6.0"
description = "Understand any codebase in 60 seconds — function-level git history analysis across 8 languages"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Brett Fox", email = "brett@foxsoftwaresolutions.com.au"},
]
keywords = ["git", "codebase", "archaeology", "analysis", "history", "ast", "tree-sitter", "hotspots", "blame", "refactoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Software Development :: Quality Assurance",
    "Environment :: Console",
]
dependencies = [
    "gitpython>=3.1.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-c>=0.23.0",
    "tree-sitter-cpp>=0.23.0",
]

[project.optional-dependencies]
embeddings = [
    "openai>=1.0.0",
    "numpy>=1.24.0",
]
config = [
    "pyyaml>=6.0",
]
all = [
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[project.scripts]
gitast = "gitast.cli:main"

[project.urls]
Homepage = "https://github.com/brettfox/gitast"
Repository = "https://github.com/brettfox/gitast"
Issues = "https://github.com/brettfox/gitast/issues"
Changelog = "https://github.com/brettfox/gitast/blob/master/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
